common/common.c et al: relocate check_perms() from upsd to common code base [#3359]
